Scripturi_Licenta_Lupei_Dacian/Aplicatie_de_control/requirements.txt
PyQt5>=5.15.0
qtawesome>=1.0.0
cryptography>=41.0.0
zenoh-python>=0.10.0